arreglo poblacion y unas funciones
tambien se añade instruccion de llave de 512 bits

+++ ClimateChangeBackend/src/main/java/com/example/ClimateChangeBackend/entities/MeasurePointsEntity.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class MeasurePointsEntity {
 
     private Long idMeasurePoints;
-    private Long latitud;
-    private Long longitud;
+    private double latitud;
+    private double longitud;
     private String sensorType;
 }
